Version 2.0 (header-only) complete
